SCREAM: moved git script utilities into their own file
The utils.py script was quite long. Shorter files are faster to
navigate.
